Fixed bug causing to fail loading operation sequence number - specific to hived resume on already filled block_log database.

+++ libraries/plugins/sql_serializer/sql_serializer.cpp
@@ -1001,7 +1001,7 @@ namespace hive
                 {
                   FC_ASSERT( data.size() == 1 );
                   const auto& record = data[0];
-                  op_sequence_id = record["_max"].as<int>();
+                  op_sequence_id = record["_max"].as<int64_t>();
                 }
                 return data_processing_status();
               }
